moves workspace user role edit to seperated component
